+++ docs/refactoring-process.md
@@ -760,6 +760,40 @@ import Icon from '@/basic/components/icons/Icon';
   - **일관성**: 모든 가격 표시가 동일한 형식 사용
   - **유지보수성**: formatPrice 함수의 명확한 역할 분리
 
+### 2.14 AdminPage 리팩토링 ✅
+
+- **목적**: 거대한 관리자 페이지 컴포넌트를 기능별로 분리하여 관심사 분리 및 재사용성 향상
+- **작업 내용**:
+  - [x] **AdminPage.tsx 대폭 간소화**:
+    - 676줄에서 44줄로 대폭 감소 (632줄 삭제)
+    - 거대한 관리자 페이지 컴포넌트를 간결한 컨테이너로 변환
+    - 단순한 레이아웃과 props 전달만 유지
+  - [x] **Admin 컴포넌트 구조 생성**:
+    - `AdminTabs.tsx`: 관리자 탭 네비게이션 및 컨테이너 (56줄)
+    - `ProductAdmin.tsx`: 상품 관리 로직 분리 (412줄)
+    - `CouponAdmin.tsx`: 쿠폰 관리 로직 분리 (286줄)
+    - `AdminSection.tsx`: 공통 관리 섹션 레이아웃 (26줄)
+  - [x] **Tabs 컴포넌트 개선**:
+    - `Tabs.tsx`: 합성 패턴 기반 탭 컴포넌트 (97줄)
+    - 제네릭 타입 지원으로 타입 안전성 확보
+    - 재사용 가능한 탭 시스템 구축
+  - [x] **관심사 분리**:
+    - 상품 관리와 쿠폰 관리 로직을 별도 컴포넌트로 분리
+    - 각 컴포넌트가 단일 책임을 가지도록 설계
+    - Props를 통한 명시적 의존성 관리
+- **검증**:
+  - [x] TypeScript 컴파일 오류 없음
+  - [x] 모든 기능 정상 작동
+  - [x] 컴포넌트 간 결합도 감소
+  - [x] 재사용성 향상
+- **개선 효과**:
+  - **코드 구조 개선**: 단일 책임 원칙(SRP) 적용
+  - **관심사 분리**: 각 관리 기능별 독립적 관리 가능
+  - **유지보수성 향상**: 변경 시 영향 범위 최소화
+  - **확장성 확보**: 새로운 관리 기능 추가 용이
+  - **재사용성**: Tabs 컴포넌트를 다른 곳에서도 사용 가능
+  - **타입 안전성**: 제네릭 타입으로 타입 안전성 확보
+
 ## 📊 현재 작업 진행 상황 업데이트
 
 | 작업                    | 상태      | 완성도 | 주요 고민사항                       |
@@ -774,6 +808,7 @@ import Icon from '@/basic/components/icons/Icon';
 | useLocalStorage 고도화  | ✅ 완료   | 100%   | 성능 vs 기능성                      |
 | App.tsx 대폭 리팩토링   | ✅ 완료   | 100%   | 거대 컴포넌트 분리 및 테스트 안정성 |
 | formatPrice 함수 수정   | ✅ 완료   | 100%   | 테스트 호환성 및 가격 표시 형식     |
+| AdminPage 리팩토링      | ✅ 완료   | 100%   | 관리자 컴포넌트 분리 및 Tabs 시스템 |
 
 ## 🎯 핵심 성과 및 인사이트
 
@@ -810,12 +845,13 @@ import Icon from '@/basic/components/icons/Icon';
 3. ✅ **Phase 2.11: useLocalStorage 훅 고도화** - 탭 간 동기화 및 성능 최적화 (완료)
 4. ✅ **Phase 2.12: App.tsx 대폭 리팩토링** - 거대 컴포넌트 분리 및 페이지 컴포넌트 분리 (완료)
 5. ✅ **Phase 2.13: formatPrice 함수 수정** - 테스트 호환성 및 가격 표시 형식 개선 (완료)
+6. ✅ **Phase 2.14: AdminPage 리팩토링** - 관리자 컴포넌트 분리 및 Tabs 시스템 구축 (완료)
 
 ### 기본과제 (남은 작업)
 
-6. **Phase 2.9: Icon 시스템 완성** - 네임스페이스 패턴 최적화 (다음 우선순위)
-7. **Phase 2.999: 긴 함수 분할** - 20줄 이하 함수 규칙 준수
-8. **Phase 3.1: 엔티티 컴포넌트와 UI 컴포넌트 분리** - README 요구사항
+7. **Phase 2.9: Icon 시스템 완성** - 네임스페이스 패턴 최적화 (다음 우선순위)
+8. **Phase 2.999: 긴 함수 분할** - 20줄 이하 함수 규칙 준수
+9. **Phase 3.1: 엔티티 컴포넌트와 UI 컴포넌트 분리** - README 요구사항
 
 ### 심화과제 (기본과제 완료 후)
 
@@ -824,4 +860,4 @@ import Icon from '@/basic/components/icons/Icon';
 
 ---
 
-**마지막 업데이트**: 2024년 (App.tsx 대폭 리팩토링, formatPrice 함수 수정, 페이지 컴포넌트 분리 완료)
+**마지막 업데이트**: 2024년 (AdminPage 리팩토링, 관리자 컴포넌트 분리, Tabs 시스템 구축 완료)
